requirements.txt
python-dotenv==1.0.1
aiogram==3.17.0
requests==2.32.3
backoff==2.2.1
ujson==5.10.0
aiohttp~=3.11.12
yarl~=1.18.3
sseclient-py==1.8.0